case insensitive matching for concepts
Todo: "match" returns true match if the BEGINNING matches, e.g., "hillary" will match to "hillaryClinton", which if we then try to use that key value to access the dataframe will fail since there is not actually a key for "hillary". Need to find an elegant way to perform exact matching, preferably avoiding adding another 'for' loop to do the job.